Add Save button to Edit Tags dialog
When editing an opening book it is very anoying if the Edit Book dialog closes
every time you add a move or change a weight, and closing the dialog was
(in WB) the only way to save the changes. A separate button now is added for
this. Pressing it will also cause a refresh of the list of book moves, so
the user can see if his change were accepted, and the consequences for the
playing percentages.