Fix grabbing of selected piece
A second static click on a selected piece should deselect it, but an attempt
to drag an already selected piece should not 'bounce off'. It was already
possible to move a selected piece to another square, but the dragging was
not animated. This is now fixed, by only skipping the code to start dragging
when the second click is part of the only-move double-click for only capture.
(Before it was skipped on any second click.)