Fix initial board sizing WB
The initial board size was based on the assumption the board would be 8x8.
This could push auxiliary windows off screen if the board was smaller,
after which EnsureOnScreen would undock them and park them in the upper-
right corner before the board assumed its true size. InitPos() is now called
before the initial sizing to et the board format correctly, InitDrawingSizes()
had to be made resitant to calling before a board size was picked for this
(because InitPos() might call that too).
  ActivateTheme() now uses the new board size, not the previous one (as was
indicated by size -2), so that -boardSize specifications inside a theme
will take effect.
  Board size petite should not be forbidden for Shogi; all pieces exist there.