Allow seting of -egtPath through menu WB
The Nalimov Path control in Common Engine Options can now also be used
to set the -egtPath; if it has the format of an egtPath string, it will
count for the latter. If both -defaultPathEGTB and -egtPath are defined
(which is really an inconsistent setting), the Nalimov path will be displayed.