Beef up variant detection
Testing whether an engine support a variant is now greatly improved.
Names that are substrings of supported variants are no longer recognized,
including base variants the engine only reports with a size-override prefix.
When the size is nonstandard, prefixes are always used on the searched name.
XBoard's New Variant dialog will refuse to close if the size overrides
(even if they specify defaults) do not correspond to a supported variant
of the first engine, but pops up an error box instead. If the problem is
that only one or more non-standard sizes are supported, this box reports
the first of them that is supported.
 GenericReadout() had to be fixed for this, as it was not setting options
that it already thought had the correct value, so that on a second try
the size-override controls were effectively ignored.