Supply oriental theme settings
Install creates a new directory .../themes/conf, in which a number of
settings files are places for defining oriental themes. These redefine the
user settings file for each of them to ~/.xboard-xxx-rc, so the settings
cannot leak into the default user settings file, but are nevertheless
persistent on revisiting the theme. XBoard now looks in this directory
for settingsfiles it could otherwise not find, so they can be invoked
with @shogi, @xq, @chu on the XBoard command line.
  Filenames starting with ~~ will now be expanded to the directory where
the XBoard data files are installed, so the installed settings files can
easily refer to installed other files (e.g. piece SVGs and textures).
The themes files will also redefine the default engine when they belong
to variants that Fairy-Max does not play.
  Some minor settings files for implementing variants that are not standard
(such as mini- and Sho Shogi) are also added in the conf directory.