Implement option complex for installing engines
Options are added to allow XBoard to be called to install engines in the
user settings files. With -addMasterOption a line can be added to the
master settings file, and this will be automatically be preceded by a
-date option to record the date it was added. This is intended for adding
-installEngine options to the master file, in order to 'broadcast' the
availability of the engine to the users.
  We now store the date in the user settings file, through a new -saveDate
option (as integer). This allows XBoard to relate the date of the saved
settings to -date stamps in the master settings file, and thus decide
if the latter (and install options that follow it) are seen for the first
time, or not. Depending on that it can decide if the install option should
be ignored.
  The volatile Boolean optio -autoClose causes XBoard to save its settings
and exit before bringing up its GUI (but after processing the command line).
This is added to facilitate calling XBoard from engine-install scripts,
with an -addMasterOption argument.