Implement -installEngine option
This volatile string option conditionally adds its argument as a line
to the engine list (-firstChessProgramNames). The condition is that the
line is not yet already there, and either that the -versionNumber option
read from settings files or command line has a value different from the
version of the binary, or that -autoClose is set. It can be used for two
purposes: to install engines from the master settings file (so that it
only works first time after upgrades, when the stored versionNumber is
obsolete), or from engine install scripts (calling XBoard with -autoClose).