Fix mate and stalemate test in Chu Shogi
Chu Shogi officially has no check, but we do want checkmate signs to
be displayed in SAN, and proper game-end messages to appear, rather than
force the user to play on until the King is captured. we also don't want
to allow him blundering away his King. So we don't use the F_IGNORE_CHECK
flag for Chu. But that means mate and stalemate will be recognized.
So it has to know stalemate is a win. Furthermore, check has to take
into account a possible spare royal. Finally, the + symbol to indicate
check should not be used in Chu, as in all Shogi variants this is reserved
to indicate promotion.