Save clocks with unfinished PGN games
When saving a game with an unfinished result, the remaining time of both
players is appended to the result comment, as "(Clocks: WHITETIME, BLACKTIME)"
with the times in seconds. When a game with unfinished result is loaded,
and the result message contains such a text, the clock settings are restored
from it accordingly. This makes it easier to resume interrupted games.