Cure weirdness when dragging outside of board
Dragging outside the board (when people maximized the window) did not
properly restore background, (leaving a trail of dragged piece). Filling
the canvas with white when it is creatd cured this. The problem most likely
was that unitialized cairo surfaces had transparancy, and writing back
transparancy doesn't helpmuch to erase things.