Put some OSX code into gtk version
Under control of #ifdef OSX code is added to integrate the GTK front-end
into OS X. This involves moving the menu bar to outside the window, and
catching the signal that OS X sends to running applications when opening
another instance was requested. Som files are renamed to conform to the
file tree of the OS X App package.