Fix name of master settings file in OS X
The name needs to be relative to the bundle path, but ~~ expansion
did not work for compiled-in defaults, as the expansion takes place
when parsing an option. The dataDir is now also kept in an array rather
than as a pointer, to allow it to appear in statically initialized tables.