Overhaul kill code
Engines were not always forcefully killed in XBoard, which could make
the tournament manager hang when an engine became unresponsive to "quit".
The various levels of killing are now revised, and the -delayAfterQuit
is absorbed in DestroyChildProcess(). There is a hard killing mode 9,
which immediately sends SIGKILL, intended for engines that have already
proven to malfunction or could not start at all. In other cases SIGTERM
is sent according to specs, but a SIGKILL is scheduled -delayAfterQuit + 1
sec later just in case the SIGTERM would be ignored.