Allow popup of TC and Common Engine from Tournament dialog
A new dialog class MasterDlg was added, similar to TransientDlg,
but to allow two dialogs to be up simultaneously. The Tournament Options
dialog was migrated to this class, so it can now get buttons to pop up
the Time Control and Common Engine dialog, like WinBoard. It even has
a button for the General options dialog.
Special care had to be taken to prevent popdown of the MasterDlg when
a Transient child is still up (similar to browser or error childs).