Allow hide/show of columns in Engine Output
A new persistent Boolean option -memoHeaders enables a new feature:
text headers are shown above the depth, score, nodes, time columns in the
engine output. Right-clicking them will subsequently hide the corresponding
data, and move the column header to above the PV in a 'not shown' list,
where it can be clicked again to re-open the column.
Configure maseter settings to switch -memoHeaders on by default.