Make Fischer castling generally available
A new volatile option -fischerCastlings allows every variant to be played
with Fischer castling, which so far was only allowed in variants FRC/CRC.
This only makes sense in shuffle variants, where King and Rooks might not
be in their orthodox locations. But even when they are, it has consequences
for the user interface, which will require dragging King on Rook as the only
way to enter castling. Like -shuffleOpenings the setting will be cleared
as soon as you switch variant (unless you switch to FRC/CRC, where they
Fischer castling is on by default). It can be set from the New Shuffle dialog.