Integrate ICS output into Chat Window
An extra pane has been added to the Chat Window to display the ICS output.
In GTK this output is colorized just as in the xterm. For Xaw this is not
possible. This pane is always visible, and the chat output can be hidden,
to double its size. Entered input will go to the selected chat (when not
hidden) as a tell or shout, or to the ICS as a command. The Chat-Partner
field is hidden together with the chat pane and the Hide button. Each of
the five chats now has a dedicated button; for unused chats (empty Chat-
Partner field) these now gets the text 'New Chat' inscribed on it.
Focussing is made more intelligent (to the Partner field when shown and
empty, otherwise to the input field).
The Up and Down arrows now work in the input field to recall old input.
 ScrollToCursor() was broken in GTK, because a mark was made for only
a single buffer. Now each buffer has its own end mark.
 Another problem was that using ScrollToCursow immediately after a memo
was shrunk (because a neighboring one expanded) it would still use the
old memo size, so that the bottom part remained out of view. Delaying it
50 ms solved that problem.
 A new option -icsUp was added to store the visibility in the settings.
It is updated only in ICS mode. The ICS Console pops up automatically
when the stored parameters say so. It is also made sticky.
 Unlike the xterm the textView does not extend background color to unwritten
positions behind the line. A separate GTK call is now added to set the
unwritten part to the same color as the background in normal text.
 A  control was added in the ICS Options dialog for setting the colors
of normal text (-colorNormal).