Add context menu to ICS console XB-GTK
A right-click handler is added to the output memos of the ICS Interaction
window. It pops up the ICS text menu. It is remembered whether the menu
was already up; if not, it will be popped down after a command is selected
from it. Like the WinBoard context menu, the commands will be able to use
the clicked word.
The -icsMenu 'command' $chat is now recognized as a special case, not to
be sent to the ICS, but executed by XBoard. It will assign one of the
five chats to the clicked name. If no empty chat is available, the last
chat will be used (but not cleared).
 The -icsMenu in the master config file is now configured to contain
this command, but unfortunately this will not become effective for users
that already have a settings file.
The Text Menu now puts $input commands in ICS Console Input field when the
ICS Input Box is not up.
 The text placed in the input field for completion (as per $input directive)
turned out to be selected in GTK, so that when you stated typing, it was
erased again!
 By making the Text menu dialog subject to WindowPlacement control,
and interrogating the poition of the ICS Console window, the Text Menu
can be popped up such that the mouse pointer is on the bottom-left button.