Also display channel tell in ICS Console during private chat
Channel tells of players for which a chat wa opened were displayed in that
chat, as well as in the channel chatbox, if the latter was open. But when
there was no channel chatbox it would not appear as colorized channel tell
in the console. This is now fixed, and in the private chat box such a
message is indicated as 'broadcast'.