Equip Board Options dialog with themes listbox
The option -themeNames is made persistent for XBoard, and its value is
displayed in a listbox added to the View -> Board dialog. The user
can select themes from this, and manipulate the contents similar to
the engine list in Load Engine. A textedit allows the user to define a
theme name for the current settings, and doing so adds the theme to the
-themeNames list.