Kludge repair of expose after startup resize
Redraw of the main board after resize is delegated to DragProc, which is
scheduled after a delay, to make sure the user is done sizing (mainly to
reposition docked windows). But at startup the program-requested resize when
an engine does not support variant normal and needs a different board format
takes so long that the finale size expose occurs after the DragProc timeout.
As a result, the board was never drawn on the resized canvas.
  This is quite hard to fix in a satisfactory way. For now the suppression
of board redrawing during resize is disabled during the first 10 seconds
after startup, as any resize during that time is likely to be the result
of a program request.