Allow writing text on pieces
A new persistent string option -inscriptions allows the user to define
characters to be written on the various piece types. The string is
interpreted as UTF-8. This option is useful for on-the-fly synthesis
of Shogi pieces, where the string can contain the various kanji in
the canonical pieceToCharTable order. Except that only a single color
has to be given.