Fix loading positions in engine-defined variant
Paste Position did not work when an engine-defined variant had redefined
the pieceToCharTable, as the position would be interpreted and sent to the
engine before receiving the setup command. Copying the pieceToCharTable
received from an earlier game to appData cures this, because it will then
be used by InitPosition, which would reset the table to default otherwise.
 Another problem was that setupSpoiledMachineBlack would not be reset
after a game with no moves, leading to a spurious 'go' command to the engine.