Fix null-move entry during play
Null moving by clicking the clock was only accepted in Edit Game and Analyze
mode, while in Chu Shogi and spin-offs it is also allowed during play.
An F_NULL_MOVE flag was introduced to make it easy to test for such variants,
and Machine White/Black mode now also interpret clock clicks as a null move
if the opponent's flag has not yet fallen.