Fix typing of null moves
Entering a null move through the move type-in could fail because the
to-square coordinates (not used in its encoding) could have invalid values,
which would reject the move in an early stage of UserMoveEvent().
Null moves are now exempted from this vaidity test, while variants where
null moves are legal now accept them in any gameMode.