Verify if font-spec looks like one in Xaw
To make XBoard resistant to mixed use of Xaw and GTK builds, which
would not understand each others font specs stored in the settings
file (they have in common), we now determine if the value looks like
an X-font. If it doesn't, the Xaw version ignores it. The GTK build already
ignored it if it did lool like an X-font spec.