Fix size of time in Engine Output window
When the thinking time exceeds an hour, the format hr:min:sec is now used,
in stead of min:sec.cents. This prevents the field from extending beyond
the tab stop when thr hours kick in, and should work upto 100 hrs.
(Going beyond that at the moment does not seem useful.)