Implement -autoInstall option
When -autoInstall is a non-empty string, XBoard will scan the system's
plugin specs directories for UCI and WB protocol, to see if new engines
were added since the last settings save. If so, they are added to the
list of installed engines.
 The variants listed in the plugin-spec files are currently ignored,
but the option is already configured to install engines for all variants.