Size seek graph to also cover board rim WinBoard
The seek graph always used the plain board size, but in WB it is
possible to diplay a rim around the board, and the seekgraph would
then be positioned in the upper-left corner of the rim, leaving the
lower and right board edge and rim uncovered. Its size is now increased
to fill the entire window (as leaving a rim around it seems not useful).