Configure XBoard for -size 49 in master settings
Now the code for auto-sizing works, it turns out this does more harm than
good in the GTK build, as it takes maximum size that fits the screen as
initial size, and the board window then cannot be shrunk below that.
So we configure a hard -size 49 for a cold startup.