Locate corner piece in presence of DarkSquares
The 'Rook' move implied by a castling indicated through an obver-sized
King step uses the corner pieces. But the corner isn't necessarily
the edge file if the board is not rectangular, but irregularly shaped
like in Omega Chess. So we have to ignore the DarkSquares, which are
not caounted as belonging to the board.