Make the promotion zone always 3 deep in Elven Chess
The promotion zone in Eleven Chess was treated as in Shogi, and set
to the board height divided by 3. It is better to always make it 3.
This only makes a difference when the boardHeight is overruled,
but Elven Chess is a very useful parent variant when a 3-deep promotion
zone is needed. (Makruk would only allow promotion to Ferz, and Grand
Chess would need holdings and allow only promotion to captured pieces.)