Allow e.p. capture on triple-Push
Some variants (like Omega Chess) have an initial triple-Push on Pawns,
wich can then e.p.-captured on both of the squares they skip. To allow
the Betza move generator to supply such e.p. captures, a bit flag is
kludged into the EP_RANK state indicator on triple pushes, while the
main value there is that of the rank directly behind the pushed Pawn.
The Betza generator then also matches the square behind it with the
e.p.-capture to-square when the falg is set.
 ApplyMove() also had to be adapted, to remove the Pawn two squares
behind the capturing one, rather than straight behind it, when this
flag is set.