Implement non-royal castling
The Betza move generator allows castlings to be specified on non-royal
pieces, and indeed the Omega-Chess 'guarding' castles Q with R. To
prevent ambiguity this is implemented as a two-leg move QxR-s (with 's'
the target square specified in the O atom). This automatically takes care
of removal of the 'Rook', so that in ApplyMove() we only have to put it
back on the proper side of the 'King'.